requirements.txt
# OCR 및 이미지 처리
openai>=1.0.0
google-cloud-vision>=3.4.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pytesseract>=0.3.10
paddleocr>=2.7.0  # PP-OCRv5 모델 지원
paddlepaddle>=2.5.0  # PaddlePaddle 프레임워크

# 딥러닝 및 AI 모델
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0

# Sibang OCR 전용
trocr>=0.1.0

# 웹 프레임워크
streamlit>=1.28.0
fastapi>=0.104.0
flask>=3.0.0  # Flask 웹 서버
uvicorn>=0.24.0
python-multipart>=0.0.6

# 유틸리티
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
aiofiles>=23.0.0

# 데이터 처리
pandas>=2.0.0


